correction /getAll ajout fonctionnalités cours+combinaison sauf /cours/combiOfCours : récupère les combinaison user/cheval pour le cours
/cheval/attr :  Ajoute un cheval à la combinaison du cours précisé

/cheval/dispo : récupère les chevaux disponible à l'heure du cours indiqué

reste aussi Token, format date

+++ back/src/main/java/Tarby_Gregoire_Web/Projet/repository/CoursRepository.java
@@ -16,7 +16,7 @@ public interface CoursRepository extends JpaRepository<Cours, Long> {
 
 	public List<Cours> findAllByDateDebutAfter(DateFormat date);
 
-	public List<Cours> findCoursByIdMoniteur(int idMoniteur);
+	public List<Cours> findCoursByIdMoniteur(long idMoniteur);
 
 	//public Cours findCoursByDateAndAndMaxcavalierAndNiveauAndTitreAndRecurrentAndIdMoniteurAndEtat(DateFormat date, int max_cavalier, int niveau, String titre, Boolean recurrent, int idMoniteur, int etat);
 	@Query("SELECT max(idCours) FROM Cours ")
